Separate checking if a PR number was found in release notes from grabbing it
Composite actions implicitly include `set -o pipefail`; this messes up lines like

`var="$( command_1 | grep pattern | command_2 )"`

which are supposed to be blank if the pattern doesn't match. The `grep` in the middle of the pipeline will not match, and exit with `1`. Then the `pipefail` kicks in and the script mysteriously fails.

I moved the testing to a non-pipefailable spot instead of slipping a `set +o pipefail` into the command substitution.